Moved migration listening to renamed DockerMigrationPlanner class.
