+++ src/main/kotlin/chain/ChainManager.kt
@@ -5,6 +5,7 @@ import communication.Message
 import communication.SyncRequest
 import communication.TransmissionType
 import data.*
+import docker.DockerMigrationPlanner
 import logging.Dashboard
 import logging.Logger
 import manager.*
@@ -26,7 +27,7 @@ class ChainManager(
     private val configuration: Configuration,
     private val vdf: VerifiableDelayFunctionManager,
     private val dht: DistributedHashTable,
-    private val docker: DockerManager,
+    private val docker: DockerMigrationPlanner,
     private val informationManager: InformationManager,
     private val blockProducer: BlockProducer
 ) {
